feat: FilterMenu component
A menu that contains all the tag filters. It will have a name search in
the future. Also adds Lucide for icons

This code is disgusting. I should be beaten with a metal rod for this.
The modal logic should be a separate component, but I couldn't figure
out how to do that with the spaghetti CSS