feat: client-side dish list
Renders the list of dishes and filters them with FilterMenu

Astro saves my butt by rendering the text immediately but I'll need to
find a way to make this interactable before the JS loads because often
users will be looking for a specific recipe and may immediately see it
before the JS loads. Perhaps React.Suspense?